Implement basic clause pruning strategy
Many changes have happened under the hood with this commit - clauses are
now stored on the heap, with the main clause list now being a vector of
clause pointers rather than a deque of clauses. This makes it much
easier to delete clauses, which I needed to do to implement pruning.